Split DefaultMeterFactory
