+++ src/libraries/Microsoft.Extensions.Diagnostics.Abstractions/src/Metrics/IMetricsSubscriptionManager.cs
@@ -1,0 +1,10 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace Microsoft.Extensions.Diagnostics.Metrics
+{
+    public interface IMetricsSubscriptionManager
+    {
+        public void Start();
+    }
+}
